requirements.txt
flask-cors-3.0.7
flask-restplus==0.12.1
pytest==4.0.2
